#1622 Move PendingFileComponents to ChatUploads folder
