+++ twake/frontend/src/app/components/ChatUploads/PendingFileComponents/styles.scss
